+++ src/main/java/org/my/companyspring/entity/product/records/TopPrdByOrdNumDTO.java
@@ -3,5 +3,5 @@ package org.my.companyspring.entity.product.records;
 public record TopPrdByOrdNumDTO(
         Long productId,
         String productName,
-        Long totalQuantity
+        Integer totalQuantity
 ) {}
